Core: Fix flag merge; some norming

+++ vm/war/init.c
@@ -6,7 +6,7 @@
 /*   By: asyed <asyed@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/05 23:42:15 by asyed             #+#    #+#             */
-/*   Updated: 2018/03/13 13:13:25 by asyed            ###   ########.fr       */
+/*   Updated: 2018/03/13 15:52:02 by sgardner         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -96,7 +96,7 @@ int			init_war(void *arena)
 			(g_taskmanager->currCycle)++;
 			(g_taskmanager->c_diecycles)++;
 		}
-		if (g_taskmanager->currCycle > *dump)
+		if (g_taskmanager->currCycle > (unsigned long)*dump)
 		{
 			dump_memory(arena);
 			return (0);
